Put mock configs in a separate file
