+++ aqueduct/gateway/tests/test_image_generation.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 
 from gateway.tests.utils import _build_chat_headers
 from gateway.tests.utils.base import GatewayIntegrationTestCase
-from gateway.tests.utils.mock_server import MockConfig
+from gateway.tests.utils.mock_configs import MockConfig
 from management.models import Request, Usage
