+++ aqueduct/gateway/tests/test_tts_stt.py
@@ -7,7 +7,7 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.test import override_settings
 
 from gateway.tests.utils.base import INTEGRATION_TEST_BACKEND, GatewayTTSSTTestCase
-from gateway.tests.utils.mock_server import MockPlainTextConfig
+from gateway.tests.utils.mock_configs import MockPlainTextConfig
 from management.models import Request
